backend/requirements.txt
# RecallIQ Version 2 - Multi-Tenant Email Reminder System
# Production-Ready Enterprise Dependencies

# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database & Caching
psycopg2-binary==2.9.10
dj-database-url==2.1.0
django-redis==5.4.0

# Background Processing
celery==5.3.4
redis==5.0.1
django_celery_beat==2.5.0
django_celery_results==2.5.0

# Security & Configuration
python-decouple==3.8
cryptography==41.0.7
django-cors-headers==4.3.1

# Web Server & Static Files
gunicorn==21.2.0
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.26.5

# Testing & Quality Assurance
coverage==7.3.2
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-xdist==3.3.1
factory-boy==3.3.0

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# System monitoring
psutil==5.9.6

# Sentry for error tracking
sentry-sdk[django]==1.38.0

# Email & OAuth Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0

# Data Processing & Export
openpyxl==3.1.2
pandas==2.3.1
pillow==10.4.0

# Additional Utilities
setuptools==68.0.0